fix duplicated slack socket test harness (#30930)
